reset to bare bone
