+++ src/pages/frens/styles.module.scss
@@ -1,0 +1,40 @@
+@import '../../styles/layout.scss';
+@import '../../styles/variables.scss';
+
+.profile {
+  display: flex;
+
+  .info {
+    display: flex;
+    flex-direction: column;
+    margin-left: 20px;
+
+    p {
+      margin-bottom: 8px;
+    }
+  }
+}
+
+.menu {
+  padding: 6px 0;
+  border-bottom: 1px solid var(--border-color);
+  transition: border-color #{$theme-duration} ease-out;
+
+  & > * {
+    margin-right: 1em;
+  }
+}
+
+.container {
+  position: relative;
+  .number {
+    position: absolute;
+    bottom: 6px;
+    right: 6px;
+    padding: 4px;
+    background-color: var(--text-color);
+    color: var(--background-color);
+    transition: background-color #{$theme-duration} ease-out,
+      color #{$theme-duration} ease-out;
+  }
+}
